requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==1.10.8
jinja2==3.1.2
python-multipart==0.0.6
